+++ test/Compiler.Tests/Emit/StaticVirtualInterfaceMembersEmitTests.cs
@@ -36,7 +36,7 @@ public class StaticVirtualInterfaceMembersEmitTests
 
             sealed interface IAdd {
                 shared {
-                    func Add(a int32, b int32) int32
+                    func Add(a int32, b int32) int32;
                 }
             }
             """;
@@ -145,7 +145,7 @@ public class StaticVirtualInterfaceMembersEmitTests
 
             sealed interface IAdd {
                 shared {
-                    func Add(a int32, b int32) int32
+                    func Add(a int32, b int32) int32;
                 }
             }
 
@@ -227,7 +227,7 @@ public class StaticVirtualInterfaceMembersEmitTests
 
             sealed interface IAdd {
                 shared {
-                    func Add(a int32, b int32) int32
+                    func Add(a int32, b int32) int32;
                 }
             }
 
@@ -296,7 +296,7 @@ public class StaticVirtualInterfaceMembersEmitTests
 
             public sealed interface IAdd {
                 shared {
-                    func Add(a int32, b int32) int32
+                    func Add(a int32, b int32) int32;
                 }
             }